+++ src/Four/Bank.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Billplz\Four;
+
+use Billplz\Three\Bank as Request;
+
+class Bank extends Request
+{
+    //
+}