+++ src/Three/Bill.php
@@ -2,99 +2,14 @@
 
 namespace Billplz\Three;
 
-use InvalidArgumentException;
+use Billplz\Base\Bill as Request;
 
 class Bill extends Request
 {
     /**
-     * Create a new bill.
+     * Version namespace.
      *
-     * @param  string  $collectionId
-     * @param  string  $email
-     * @param  string  $mobile
-     * @param  string  $name
-     * @param  \Money\Money|int  $amount
-     * @param  array|string  $callbackUrl
-     * @param  string  $description
-     * @param  array  $optional
-     *
-     * @throws \InvalidArgumentException
-     *
-     * @return \Laravie\Codex\Response
+     * @var string
      */
-    public function create(
-        $collectionId,
-        $email,
-        $mobile,
-        $name,
-        $amount,
-        $callbackUrl,
-        $description,
-        array $optional = []
-    ) {
-        if (empty($email) && empty($mobile)) {
-            throw new InvalidArgumentException('Either $email or $mobile should be present');
-        }
-
-        $body = array_merge(
-            compact('email', 'mobile', 'name', 'amount', 'description'),
-            $optional
-        );
-
-        $body['collection_id'] = $collectionId;
-
-        $body = $this->parseRedirectAndCallbackUrlFromRequest($body, $callbackUrl);
-
-        return $this->send('POST', 'bills', [], $body);
-    }
-
-    /**
-     * Show an existing bill.
-     *
-     * @param  string  $id
-     *
-     * @return \Laravie\Codex\Response
-     */
-    public function show($id)
-    {
-        return $this->send('GET', "bills/{$id}");
-    }
-
-    /**
-     * Show an existing bill transactions.
-     *
-     * @param  string  $id
-     * @param  array   $optional
-     *
-     * @return \Laravie\Codex\Response
-     */
-    public function transaction($id, array $optional = [])
-    {
-        return $this->client->resource('Bill.Transaction', $this->getVersion())
-                    ->show($id, $optional);
-    }
-
-    /**
-     * Destroy an existing bill.
-     *
-     * @param  string  $id
-     *
-     * @return \Laravie\Codex\Response
-     */
-    public function destroy($id)
-    {
-        return $this->send('DELETE', "bills/{$id}");
-    }
-
-    /**
-     * Parse webhook data for a bill.
-     *
-     * @param  array  $data
-     *
-     * @return array
-     */
-    public function webhook(array $data = [])
-    {
-        return $this->sanitizeTo($data);
-    }
+    protected $version = 'v3';
 }