+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This changelog references the relevant changes (bug and security fixes) done to `jomweb/billplz`.
 
+## 1.0.4
+
+Released: 2017-10-30
+
+### Added
+
+* Add `Billplze\Three\Bank::checkAccount()` to handle checking account status as prefered alternative instead of `Billplz\Three\Check::bankAccount()`.
+* Add `Billplz\Three\Bank::supportedForFpx()` to check list of banks support with FPX. ([@mrkaymy](https://github.com/mrkaymy))
+* Add X-Signature verification for `Billplz\Three\Bill::webhook()` which will throws `Billplz\Exceptions\FailedSignatureVerification` exception if the hash is not equals.
+
+## 1.0.3
+
+Released: 2017-10-12
+
+### Changes
+
+* Allows to cast to `Duit\MYR`.
+
 ## 1.0.2
 
 Released: 2017-10-08